+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Perlock Changelog
 ==================
 
+Changes in version 0.3.0 (2014-10-06)
+-------------------------------------
+
+* Added new interface `LifecycleListener` to allow for notification of important lifecycle events.
+* Improved robustness of the default path watcher implementation
+* Added file system tests using [Jimfs](https://github.com/google/jimfs)
+
 Changes in version 0.2.0 (2014-02-21)
 -------------------------------------